# backend/school-management/public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Management API",
        "description": "Documentación de la API para el sistema de gestión escolar",
        "contact": {
            "name": "Equipo de desarrollo CBTa71",
            "email": "soporte@cbta71.edu.mx"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000/api",
            "description": "Servidor local"
        },
        {
            "url": "https://nginx-production-728f.up.railway.app/api",
            "description": "Servidor de producción"
        }
    ],
    "paths": {
...
        "/api/v1/careers": {
            "get": {
                "tags": [
                    "Careers"
                ],
                "summary": "Obtener todas las carreras",
                "operationId": "47f4386bfc85fc8263fc82d76256e604",
                "parameters": [
                    {
                        "name": "forceRefresh",
                        "in": "query",
                        "description": "Forzar actualización del caché (true o false).",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de carreras",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "careers": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/DomainCareer"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Carreras encontradas."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1//careers/{id}": {
            "get": {
                "tags": [
                    "Careers"
                ],
                "summary": "Obtener una carrera por ID",
                "operationId": "1da3b0431b893f26ba50522ac3fb75db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la carrera",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "forceRefresh",
                        "in": "query",
                        "description": "Forzar actualización del caché (true o false).",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carrera encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "career": {
                                                    "$ref": "#/components/schemas/DomainCareer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Carrera encontrada."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Careers"
                ],
                "summary": "Eliminar una carrera",
                "operationId": "c9604f40880663bcf210b1cd508ea16c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la carrera a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carrera eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Carrera eliminada con éxito."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Careers"
                ],
                "summary": "Actualizar una carrera existente",
                "operationId": "4dab81598c9454d2d4242a634e909705",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la carrera",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "career_name": {
                                        "type": "string",
                                        "example": "Ingeniería en Software"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Carrera actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "updated": {
                                                    "$ref": "#/components/schemas/DomainCareer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Carrera actualizada."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1//careers": {
            "post": {
                "tags": [
                    "Careers"
                ],
                "summary": "Crear una nueva carrera",
                "operationId": "87edb3b595b3cb81cc8d693a736c3042",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DomainCareer"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Carrera creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "career": {
                                                    "$ref": "#/components/schemas/DomainCareer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Carrera creada."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DomainCareer": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "career_name": {
                        "type": "string",
                        "example": "Matematicas"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}